Clean up the parse_line method
This is not very robust and the regex will need to be cleaned up to
handle situations like 261 where there is not a comma between console
and year, or 76 where we have a year like "2007/2000".

This also returns an array of consoles rather than a singular console,
so if you want just a singular console, just call flatten on the array.